requirements.txt
spotipy
python-dotenv
requests
tqdm
# streamlit
# streamlit-card
Flask
flask_caching